Refactor progress package
