+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/progress/navigation/ProgressScreen.kt
@@ -1,9 +1,9 @@
-package org.hyperskill.app.android.track_progress.navigation
+package org.hyperskill.app.android.progress.navigation
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentFactory
 import com.github.terrakok.cicerone.androidx.FragmentScreen
-import org.hyperskill.app.android.track_progress.fragment.ProgressFragment
+import org.hyperskill.app.android.progress.fragment.ProgressFragment
 
 object ProgressScreen : FragmentScreen {
     override fun createFragment(factory: FragmentFactory): Fragment =
