+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/progress/ui/ProgressDefaults.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.android.track_progress.ui
+package org.hyperskill.app.android.progress.ui
 
 import androidx.compose.ui.unit.dp
