+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/progress/ui/project/ProjectProgressSkeleton.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.android.track_progress.ui.project
+package org.hyperskill.app.android.progress.ui.project
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -12,8 +12,8 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.tooling.preview.Preview
 import org.hyperskill.app.android.R
 import org.hyperskill.app.android.core.view.ui.widget.compose.ShimmerLoading
-import org.hyperskill.app.android.track_progress.ui.BlockHeaderSkeleton
-import org.hyperskill.app.android.track_progress.ui.ProgressDefaults
+import org.hyperskill.app.android.progress.ui.BlockHeaderSkeleton
+import org.hyperskill.app.android.progress.ui.ProgressDefaults
 
 @Composable
 fun ProjectProgressSkeleton(
