+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/progress/ui/track/TrackProgressContent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.android.track_progress.ui.track
+package org.hyperskill.app.android.progress.ui.track
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -11,9 +11,9 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import org.hyperskill.app.android.R
-import org.hyperskill.app.android.track_progress.ui.BlockHeader
-import org.hyperskill.app.android.track_progress.ui.GeneralStatistics
-import org.hyperskill.app.android.track_progress.ui.ProgressDefaults
+import org.hyperskill.app.android.progress.ui.BlockHeader
+import org.hyperskill.app.android.progress.ui.GeneralStatistics
+import org.hyperskill.app.android.progress.ui.ProgressDefaults
 import org.hyperskill.app.progresses.view.ProgressScreenViewState
 
 @Composable
